TST/DOC: review follow-ups for the read_csv CPU-limit clamp
Move the pandas/compat/_cpu.py unit tests out of the io/parser test module
into pandas/tests/test_cpu.py, mirroring test_optional_dependency.py.  Cover
the cases the first pass missed: the cgroup v1 -1 unlimited sentinel, a
sub-1-CPU quota (docker --cpus=0.5, Kubernetes 500m), and _read_sysfs_int
against a real non-numeric file.

test_parallel_default_off_on_windows pinned os.cpu_count but not
available_cpu_count, so it would fail on any runner with a single usable CPU
-- exactly the environment this branch targets.

Doc corrections: io.rst listed the eligibility conditions as sufficient, but a
process limited to one CPU now satisfies all of them and still reads serially;
the 5 MB threshold applies to the data section, not the whole file; and the
mode.max_threads option doc conflated Windows (where an explicit setting still
re-enables parallel reading) with Emscripten (where it is ignored).
