+++ pandas/core/config_init.py
@@ -463,10 +463,10 @@ max_threads_doc = """
     Maximum number of worker threads for parallel operations (e.g. ``read_csv``
     for large files).  ``None`` (the default) means use ``min(os.cpu_count(), 4)``,
     further limited to the CPUs available to the process (CPU affinity and cgroup
-    limits); on Windows and WASM the default is ``1``, as parallel reading is
-    disabled there.  Set to ``1`` to disable parallel execution, or to a fixed
-    number to raise the cap or to limit thread usage when pandas is embedded in a
-    larger parallel workflow.
+    limits); on Windows the default is ``1``, as parallel reading is not faster
+    there.  Set to ``1`` to disable parallel execution, or to a fixed number to
+    raise the cap or to limit thread usage when pandas is embedded in a larger
+    parallel workflow.  Ignored on Emscripten/Pyodide, which cannot spawn threads.
 """
