+++ pandas/io/parsers/readers.py
@@ -376,12 +376,12 @@ def _default_n_workers() -> int:
     """
     Default worker count for a parallel ``read_csv``.
 
-    ``mode.max_threads`` wins whenever it is set (except on WASM, which cannot
-    spawn threads at all).  Otherwise parallel reading is off by default on
-    Windows, and elsewhere defaults to ``_MAX_DEFAULT_WORKERS`` logical CPUs,
-    clamped to the CPUs actually available to the process (CPU affinity /
-    cgroup limits) so that an embedded or containerised pandas does not
-    oversubscribe its allocation.
+    ``mode.max_threads`` wins whenever it is set (except on Emscripten, which
+    cannot spawn threads at all).  Otherwise parallel reading is off by default
+    on Windows, and elsewhere is the smallest of the machine's logical CPU
+    count, ``_MAX_DEFAULT_WORKERS``, and the CPUs actually available to the
+    process (CPU affinity / cgroup limits) -- so that an embedded or
+    containerised pandas does not oversubscribe its allocation.
     """
     max_threads = get_option("mode.max_threads")
     if sys.platform == "emscripten":
@@ -447,7 +447,8 @@ def _can_parallelize_csv(filepath_or_buffer, kwds: dict) -> bool:
       path, and that error must not be masked.
     * ``on_bad_lines`` is not ``"warn"`` - chunk workers would report
       chunk-relative (i.e. wrong) line numbers.
-    * The file is at least ``_PARALLEL_READ_MIN_BYTES`` bytes large.
+    * Both the file and its data section (i.e. excluding the header preamble)
+      are at least ``_PARALLEL_READ_MIN_BYTES`` bytes large.
 
     Note that a chunk boundary landing on a newline embedded inside a quoted
     field leaves that chunk's parser inside an open quote at EOF, which raises
